aads-stats-v2/requirements.txt
supabase-py==1.0.0
python-dotenv==1.0.0
